+++ samples/node/index.js
@@ -2,8 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import platform from "../platform/webext/index.js";
-import base from "./base.js";
+"use strict";
 
-export { ErrorType } from "../core/error/error_type.js";
-export default base(platform);
+import Glean from "@mozilla/glean/node";
+
+import main from "./main.js";
+
+main();
+// Wait for Glean to finish up before exiting the app.
+await Glean.shutdown();
+