+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean.js/compare/v0.19.0...main)
+[Full changelog](https://github.com/mozilla/glean.js/compare/v0.20.0...main)
+
+# v0.20.0 (2021-09-17)
+
+[Full changelog](https://github.com/mozilla/glean.js/compare/v0.19.0...v0.20.0)
+
+* [#696](https://github.com/mozilla/glean.js/pull/696): Expose Node.js entry point `@mozilla/glean/node`.
+* [#695](https://github.com/mozilla/glean.js/pull/695): Implement PlatformInfo module for the Node.js platform.
+* [#695](https://github.com/mozilla/glean.js/pull/730): Implement Uploader module for the Node.js platform.
+
 
 # v0.19.0 (2021-09-03)